Moved topo and elevationMap to points
