+++ app/src/main/java/com/github/bgabriel998/softwaredevproject/points/ElevationMap.java
@@ -1,11 +1,8 @@
-package com.github.bgabriel998.softwaredevproject.map;
+package com.github.bgabriel998.softwaredevproject.points;
 
 import android.util.Log;
 
 import androidx.core.util.Pair;
-
-import com.github.bgabriel998.softwaredevproject.points.POIPoint;
-import com.github.bgabriel998.softwaredevproject.points.UserPoint;
 
 import org.osmdroid.util.BoundingBox;
